fix(docs): Move Sphinx documentation source to docs directory
Updated the .readthedocs.yaml configuration to point to the new
location of the Sphinx source files in the docs directory. Removed
obsolete Sphinx make.bat and Makefile, along with other related
files, to streamline the documentation structure. This change
supports the transition to Read the Docs for better versioning and
management of documentation.